Create ko.ini
As I have mentioned in the source file, The Korean font file unfortunately fails to show all 11,772 korean letters. However, since I'm not savvy in the field of programming, I'm leaving the downlolad link of the free Korean font(Nanum-Gothic), with a binding hope that one of the developers implement the Korean fonts into Stepmania.